[FEAT] NicknameCheck관련 로직 interface분리 및 extension으로 기본구현(#17)
