+++ Plu-iOS/Plu-iOS/Scenes/Onboarding/MangerImpl/OnboardingManagerStub.swift
@@ -10,7 +10,7 @@ import Foundation
 final class OnboardingManagerStub: NicknameManager {
     private let storedNicNames = ["의성", "민재", "찬미"]
     
-    func judgeInputNicknameVaild(input: String) async throws -> Bool {
+    func inputNicknameIsVaild(input: String) async throws -> Bool {
         return !storedNicNames.contains(input)
     }
 }
